Fix Grid warning on first load after cache-rebuild
This declares the PHP-generated Grid Config as an external ressource (which it kind of is: it only works via PHP, and is thus not available to Drupal via the filesystem).

The Locale module, responsible for the error, doesn't try to parse/load external files, so the file is ignored.

But the browser on the other hand loads it just as usual.